app/requirements.txt
dash==1.19.0
dash-bootstrap-components==0.11.2
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
DateTime==4.3
Flask==1.1.2
Flask-Compress==1.8.0
gunicorn==20.0.4
matplotlib==3.5.1
nbformat==5.1.3
numpy==1.21.4
openpyxl==3.0.9
pandas_datareader==0.10.0
pandas==1.3.5
pathlib==1.0.1
plotly==5.5.0
python-dateutil==2.8.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
SQLAlchemy==1.4.28
statsmodels==0.13.1
tqdm==4.62.3
xgboost==1.5.1
zipfile39==0.0.4.2